[svn-r196] Changes since 19980128 ----------------------
./MANIFEST
	Added new files.

./bin/Makefile.dist		[NEW]
	This file becomes the default top-level Makefile which invokes
	configure with no arguments and then re-runs make.  It's for
	those people that don't read the INSTALL file.

./bin/release			[NEW]
	Build HDF release tarballs.  Arguments are a list of tarball
	types and are: tar, compress, gzip, or bzip2.  The default is
	gzip. If the only argument is `all' then all types of tarballs
	are created.  The results are put in the `releases' directory.
	Make sure you keep MANIFEST up to date.  This program requires
	perl-5.003 or better.

./src/H5AC.c
	Fixed a printf format.

+++ src/H5AC.c
@@ -873,12 +873,13 @@ H5AC_debug(H5F_t *f)
         } else {
             sprintf(ascii, "%7.2f%%", miss_rate);
         }
-        fprintf(stderr, "   %18s: %8d %8d %7s %8d%+-9d\n", s,
+        fprintf(stderr, "   %18s: %8u %8u %7s %8u%+-9ld\n", s,
                 cache->diagnostics[i].nhits,
                 cache->diagnostics[i].nmisses,
                 ascii,
                 cache->diagnostics[i].ninits,
-             cache->diagnostics[i].nflushes - cache->diagnostics[i].ninits);
+		((long)(cache->diagnostics[i].nflushes) -
+		 (long)(cache->diagnostics[i].ninits)));
     }
 
     FUNC_LEAVE(SUCCEED);
